+++ CHANGELOG.md
@@ -7,7 +7,14 @@
 #### Fixes
 
 * Your contribution here.
-* [#804](https://github.com/ruby-grape/grape-swagger/pull/804): Don't overwrite model description with the route description. - [@Bhacaz](https://github.com/Bhacaz)
+
+
+### 1.3.0 (September 3, 2020)
+
+#### Features
+
+* [#804](https://github.com/ruby-grape/grape-swagger/pull/804): Don't overwrite model description with the route description - [@Bhacaz](https://github.com/Bhacaz).
+
 
 ### 1.2.1 (July 15, 2020)